added some example code
git-svn-id: http://svn.newfolderconsulting.com/prt@133 0c11b213-6b92-4ae5-b10c-e53e59125080

+++ classifiers/prtClassDlrt.m
@@ -57,7 +57,7 @@ classdef prtClassDlrt < prtClass
             n = TestDataSet.nObservations;
 
             uClasses = Obj.DataSet.uniqueClasses;
-            classCounts = histc(Obj.DataSet.targets,uClasses);
+            classCounts = histc(double(Obj.DataSet.targets),double(uClasses));
             n0 = classCounts(1);
             n1 = classCounts(2);
